Added a dice roll, party, and more weather commands
The dice roll command simply returns a random number, the party command outputs emojis, and the additional open weather map commands include the high and low temperatures for the day, when the sun rises/sets, the time the information was received, and other scientific information. Fixed an issues where the weather command wouldn't include spaces or punctuation. Also added a zip code condition for the main weather function.